+++ packages/metrics/chain.go
@@ -1,0 +1,11 @@
+package metrics
+
+// type StateManagerMetrics interface {
+// }
+//
+// type ConsensusMetrics interface {
+// }
+
+type ChainMetrics interface {
+	MempoolMetrics
+}
